==> scripts/01_prep_gam_predictors.Rmd <==
---
editor_options: 
  chunk_output_type: console
---

# Prepare Climate Data for Interpolation

```{r}
# data libs
# use tidyverse because small data
# and collab readability
library(tidyverse) # because data are small
library(lubridate)

# spatial libs
library(terra)
library(sf)
```

## Prepare climate data

### Read data and add date

```{r}
clim_files = list.files("data/raw/iitg-climate", full.names = T)
data = map(clim_files, read_delim, 
           delim = " ", col_names = c("ppt", "tmin", "tmax", "wind")
)

# seq dates
seq_date = seq(
  ymd("1870-01-01"), 
  ymd("2018-12-31"), 
  by = "days"
)

# add date, year, month, and season
data = map2(
  data, clim_files,
  function(df, clf) {
    
    # get the coordinates
    coords = str_extract_all(clf, "\\d+\\.\\d+")[[1]]
    mutate(
      df, 
      # add the date, year, and month, and season
      # dry season is december to may 12 - 5
      date = seq_date,
      year = year(date),
      month = month(date),
      season = if_else(between(month, 6, 11), "rainy", "dry"),
      
      # convert coordinates
      x = as.numeric(coords[2]),
      y = as.numeric(coords[1])
    )
  }
)

# bind all data
data = bind_rows(data)

# check data
assertthat::assert_that(
  nrow(
    distinct(data, x, y)
  ) == length(clim_files),
  msg = "error in some weather station coordinates"
)

# filter for southern western ghats
# y < 12 and y > 9
data = filter(
  data,
  between(y, 9, 12)
)

# save data
write_csv(
  data,
  file = "data/output/data_clim_daily.csv"
)
```

### Summarise by month

```{r}
# get daily mean temperature
data = mutate(
  data,
  tmean = (tmax + tmin) / 2
)

# prepare to summarise
data = group_by(
  data,
  x, y, month, year, season,
) |> 
  # groups will be added by default
  select(
    ppt, tmean
  )

# summarise monthly ppt sum and tmean mean and sd
data = summarise(
  data,
  ppt = sum(ppt),
  t_mean = mean(tmean),
  t_sd = sd(tmean)
)

# save summary
write_csv(
  data,
  "data/output/data_monthly_climate.csv"
)
```

### Prepare for spatial data

```{r}
# get unique data, i.e., from unique coords
coords = ungroup(data) |> 
  distinct(x, y) |> 
  mutate_all(
    as.numeric
  )
```

## Prepare spatial data

### Prepare elevation data

```{r}
# get unique coordinates
coord_sf = st_as_sf(
  coords, 
  coords = c('x', 'y'), 
  crs = 4326
)

# get extent to clip elevation
ext = st_bbox(coord_sf) |> 
    st_as_sfc() |> 
    st_transform(32643) |>
    st_buffer(25000) |>
    st_transform(4326)

# load elevation raster
elevation = terra::rast('data/spatial/Elevation/alt')
# crop by extent of study area
elevation_hills = terra::crop(elevation, as(ext, 'Spatial'))

# save a smaller piece
terra::writeRaster(
  elevation_hills, 
  'data/spatial/raster_elevation.tif',
  overwrite = TRUE
)
```

### Prepare coastline data

```{r}
# load coastline and elevation
coast = st_read("data/spatial/India/India_Boundary.shp")
coast_line = st_cast(coast, "MULTILINESTRING")

# load small elevation raster
elev = terra::rast("data/spatial/raster_elevation.tif")
elev_utm = terra::project(elev, "epsg:32643")
```

Resample raster to 600 m.

```{r}
# resample the raster to 600 m from 30m
elev_utm_600 = rast(ext(elev_utm), resolution = 600)
elev_utm_600 = terra::resample(
  elev_utm, elev_utm_600
)
crs(elev_utm_600) = "epsg:32643"

# recast to 4326
elev_600 = terra::project(elev_utm_600, "epsg:4326")
```

```{r}
# make coast raster
coast_raster = terra::distance(elev_600, terra::vect(coast_line))
coast_raster = terra::mask(coast_raster, terra::vect(coast))
```

```{r}
# latitude raster using copy of elevation and assigning coordinate
lat_raster = copy(elev_600)
terra::values(lat_raster) = terra::yFromCell(
  lat_raster, 
  cell = seq(length(values(lat_raster)))
)
```

```{r}
# stack rasters
stack = c(coast_raster, elev_600, lat_raster)
names(stack) = c("coast", "elev", "lat")
```

```{r}
terra::writeRaster(
  stack,
  "data/output/raster_gam_stack.tif",
  overwrite = T
)
```

## Link predictors to climate data

```{r}
# extract covariates using high-res raster
coords = mutate(
  coords,
  
  # extract elevation
  elev = terra::extract(
    elevation_hills,
    st_coordinates(coord_sf)
  )$alt,
  
  # get distance to coastline
  coast = as.numeric(
    st_distance(coord_sf, coast_line)
  ),
  
  # get latitude
  lat = y
)

# link to climate data summary
data = left_join(
  data,
  coords
)

# bin into 10 year intervals
data = mutate(
  data,
  year_bin = floor(year / 10) * 10
)

# save data
write_csv(
  data,
  "data/output/data_for_gam.csv"
)
```
