==> scripts/02_prep_gam_comparisons.Rmd <==
---
editor_options: 
  chunk_output_type: console
---

# Validating semi-Spatial GAMs for climate

## Prepare weather and spatial data

```{r}
# for data
library(tidyverse)
library(sf)
library(readxl)

# for gam
library(mgcv)

# for rasters
library(terra)
```

Load data prepared for the GAM, and the stack predictor raster.

```{r}
# read data
data = read_csv("data/output/data_for_gam.csv")
stack = terra::rast("data/output/raster_gam_stack.tif")
```

Prepare data covariates and raster.

```{r}
# prepare data for gam fitting
data = mutate(
  data,
  
  # divide distance by 1000 for km
  coast = coast / 1000,
  
  # add some error to all physical variables
  # hopefully prevents model issues
  coast = coast + rnorm(length(coast), 0.1, sd = 0.01),
  elev = elev + rnorm(length(coast), 1, sd = 0.01),
  lat = lat + rnorm(length(lat), 0, sd = 0.01)
)

# set distance to coast raster in km
values(stack[["coast"]]) = values(stack[["coast"]]) / 1000
```

Pivot longer to separate by variable.

```{r}
# pivot longer
data = pivot_longer(
    data,
    cols = c("ppt", "t_mean", "t_sd"),
    names_to = "climvar"
)
data = dplyr::select(data, !month)

# select period 1970 -- 2010 where we also have satellite data
data = filter(data, between(year, 1970, 2010))
```

Split the data by variable.

```{r}
data = nest(
  data, data = !c(climvar, season, year_bin)
)
```

## Fit candidate models

Prepare model formulas.

```{r}
# elevation with 3 knots
form_elev_model = value ~ s(elev, k = 3)

# elevation and coast
form_elev_coast = value ~ s(elev, k = 3) + coast + lat

# elevation and coast with lat
form_elev_colat = value ~ s(elev, k = 3) + s(coast, lat, k = 5)
```

```{r}
# make combinations
data = crossing(
  data, 
  forms = c(form_elev_model, form_elev_coast, form_elev_colat)
)
```

```{r}
# fit models
data = mutate(
  data,
  mod = map2(data, forms, function(df, form) {
    gam(
      formula = form,
      data = df
    )
  })
)
```

Save model data.

```{r}
save(data, file = "data/output/model_gam_climate.Rds")
```

## Predict over static environment

Get predictions

```{r}
# save model predictions
model_pred = map(
  data$mod, function(g) {
    predict(stack, g, type = "response") # order matters here
  }
)

model_pred = Reduce(model_pred, f = c)
terra::writeRaster(model_pred, "data/output/data_gam_pred.tif")
```

```{r}
model_pred = terra::rast("data/output/data_gam_pred.tif")
```

Average predictions over 1970 -- 2010.

```{r}
# assign as list column
data = mutate(
  data,
  pred = as.list(model_pred)
)

# summarise rasters using a reduce over the list
# basically gets the average over 1970 -- 2010
data_pred = ungroup(data) %>% 
  group_by(season, climvar, forms) %>% 
  summarise(
    mean_pred = list(Reduce(pred, f = mean))
  )

# make raster stack
gam_pred_avg = data_pred$mean_pred
gam_pred_avg = Reduce(f = c, gam_pred_avg)

# save averaged predictions
terra::writeRaster(gam_pred_avg, filename = "data/output/gam_pred_valid_avg.tif")
```

```{r}
# load saved pred
gam_pred_avg = terra::rast("data/output/gam_pred_valid_avg.tif")

# write data for linking
data_pred |>
  
  # ungroup the data
  ungroup() |>
  
  # get unique season, variable, and model formula
  distinct(season, climvar, forms) |>
  
  # then get the model formula as a character
  mutate(
    forms = map_chr(forms, function(l) {
      str_flatten(as.character(l)[c(2, 1, 3)])
    })
  ) |>
  
  # save to file
  write_csv(
    file = "data/output/gam_model_formulas.csv"
  )
```

## Get BIOCLIM data

### Get mean temperature

```{r}
patterns <- c("tmin", "tmax")

# list the filepaths
tkAvg <- map(patterns, function(pattern) {
  # list the paths
  files <- list.files(
    path = "data/spatial/chelsa",
    full.names = TRUE,
    recursive = TRUE,
    pattern = pattern
  )
})

# now run over the paths and read as rasters and crop by buffer
tkAvg <- map(tkAvg, function(paths) {
  # going over the file paths, read them in as rasters, convert CRS and crop
  tempData <- map(paths, function(path) {
    a <- terra::rast(path)
    a <- terra::crop(a, terra::ext(stack))
    a
  })
  # convert each to kelvin, first dividing by 10 to get celsius
  tempData <- map(tempData, function(tmpRaster) {
    tmpRaster <- (tmpRaster / 10) + 273.15
    tmpRaster
  })
})

# recursively get the mean temperature for each month
# assign names
names(tkAvg) <- patterns

# go over the tmin and tmax and get the average monthly temp
tkAvg <- map2(tkAvg[["tmin"]], tkAvg[["tmax"]], function(tmin, tmax) {
  # return the mean of the corresponding tmin and tmax
  # still in kelvin
  terra::mean(c(tmin, tmax))
})

# check temp data
assertthat::assert_that(
  length(tkAvg) == 12,
  msg = "temp raster list has fewer than 12 months"
)

# assign names
names(tkAvg) = sprintf("month_%i", seq(12))

# separate rainy and dry season
temp_rainy = Reduce(tkAvg[seq(6, 11)], f = `c`) |>
  terra::mean()
temp_dry = Reduce(tkAvg[c(12, seq(5))], f = `c`) |>
  terra::mean()

# convert values back to celsius
chelsa_t_mean = c(temp_rainy, temp_dry) - 273.15
names(chelsa_t_mean) = c("chelsa_temp_rainy_6_11", "chelsa_temp_dry_12_5")

# save stack
terra::writeRaster(
  chelsa_t_mean,
  filename = "data/output/chelsa_temp_stack.tif",
  overwrite = TRUE
)
```

### Get precipitation

```{r}
# list precipitation rasters --- get from BIOCLIM
ppt <- list.files(
  path = "data/spatial/chelsa",
  full.names = TRUE,
  recursive = TRUE,
  pattern = "prec"
)

# now run over the paths and read as rasters and crop by buffer
ppt <- map(ppt, function(path) {
  # going over the file paths, read them in as rasts
  a <- terra::rast(path)
  terra::crop(a, terra::ext(stack))
})

# separate rainy and dry season
ppt_rainy = Reduce(ppt[seq(6, 11)], f = `c`) |>
  terra::mean()
ppt_dry = Reduce(ppt[c(12, seq(5))], f = `c`) |>
  terra::mean()

# make and save stack
chelsa_ppt_sum = c(ppt_rainy, ppt_dry)
names(chelsa_ppt_sum) = c("chelsa_ppt_rainy_6_11", "chelsa_ppt_dry_12_5")

# save stack
terra::writeRaster(
  chelsa_ppt_sum,
  filename = "data/output/chelsa_ppt_stack.tif",
  overwrite = TRUE
)
```

## Get sample coordinates

Select some coordinates and save data.

```{r}
# get coordinates from terra
coords = terra::xyFromCell(
  stack, 
  cell = seq(length(values(stack[[1]])))
)

# sample 10 groups of 1000
coords = coords[sample(1e4, replace = FALSE), ] |>
  as_tibble()

# extract data at locations
sample_locations =
  mutate(
  coords,
  terra::extract(stack, coords)
) |>
    mutate(
      terra::extract(chelsa_ppt_sum, coords)
    ) |>
    mutate(
      terra::extract(chelsa_t_mean, coords)
    )
```

```{r}
# save data
write_csv(
  sample_locations,
  file = "data/output/data_sample_coords_gam_validation.csv"
)
```

```{r}
sample_locations = read_csv(
  "data/output/data_sample_coords_gam_validation.csv"
)
```

## Compare BIOCLIM and model predictions

### Link with samples

First sample locations.

```{r}
# melt data
sample_locations = 
  sample_locations |>
  pivot_longer(
    cols = !c("x", "y", "ID", "coast", "elev", "lat")
  )

# assign season and climvar
sample_locations = mutate(
  sample_locations,
  season = str_extract(
    name, pattern = "dry|rainy"
  ),
  climvar = str_extract(
    name, pattern = "temp|ppt"
  ),
  climvar = if_else(
    climvar == "temp", "t_mean", "ppt"
  )
)
```

Link with GAM predictions.

```{r}
# remove t_sd
data_pred = filter(
  data_pred,
  climvar != "t_sd"
)

# nest by season and variable
sample_locations = nest(
  sample_locations,
  chelsa = !c("season", "climvar")
)

# link with data predictions
data_pred = left_join(
  data_pred,
  sample_locations
)
```

Get GAM predictions at chelsa sample locations.

```{r}
# sample locations from GAM predictions
data_pred = mutate(
  data_pred,
  chelsa = map2(chelsa, mean_pred, function(ch, pr) {
    ch |>
      rename(
        bioclim_val = "value"
      ) |>
      mutate(
        pred_val = terra::extract(pr, ch[, c("x", "y")]) 
        # the naming doesn't quite work
      )
  })
)

# make copy
data_gam_validate = data_pred |>
  select(season, climvar, forms, chelsa)

# unnest data
data_gam_validate = unnest(
  data_gam_validate,
  cols = chelsa
)

# convert from formula to characters
data_gam_validate = mutate(
  data_gam_validate,
  forms = as.character(forms)
)

## extract pred val
data_gam_validate = data_gam_validate |>
  mutate(
    pred_val = pred_val$lyr1
  )

# rename and drop some variables
# save data
write_csv(
  data_gam_validate,
  file = "data/output/data_gam_validate_compare.csv"
)
```


```{r}
# use mean absolute error Measure for 10 chunks of 1000 coordinates
data_gam_validate = group_by(
  data_gam_validate,
  season, climvar, forms
) |>
  mutate(
  group = rep(seq(10), each = 1e3L)
)

# save mean absolute error measures for plotting - plot script in figure scripts
data_mae = 
  group_by(
    data_gam_validate,
    season, climvar, forms, group
  ) |>
  summarise(
    mae = mean(
      abs(
        bioclim_val - pred_val
      ), na.rm = TRUE
    )
  )

# save data
write_csv(
  data_mae,
  file = "data/output/data_gam_comparison_mae.csv"
)
```

## GAM predictions at survey sites

```{r}
# read survey sites
survey_sites = read_excel(
  "data/raw/2020-resurvey-sites-master.csv.xlsx"
) |>
  rename(
    x = "longitude",
    y = "latitude"
  )

# sample rasters at survey sites
# BIOCLIM samples
survey_sites = mutate(
  survey_sites,
  terra::extract(
    c(chelsa_t_mean, chelsa_ppt_sum),
    survey_sites[, c("x", "y")]
  )
)

# remove id
survey_sites = select(survey_sites, -ID)

# melt data
survey_sites = pivot_longer(
  survey_sites,
  cols = !c("site_code", "x", "y")
) |>
  mutate(
    season = str_extract(
    name, pattern = "dry|rainy"
  ),
  climvar = str_extract(
    name, pattern = "temp|ppt"
  ),
  climvar = if_else(
    climvar == "temp", "t_mean", "ppt"
  )
)
# nest data
survey_sites = nest(
  survey_sites,
  chelsa = !c("season", "climvar")
)

# remove chelsa data
data_pred = select(data_pred, -chelsa)

# link survey data
survey_sites = left_join(
  survey_sites,
  data_pred
)

survey_sites = mutate(
  survey_sites,
  chelsa = map2(chelsa, mean_pred, function(ch, pr) {
    ch |>
      rename(
        bioclim_val = "value"
      ) |>
      mutate(
        pred_val = terra::extract(pr, ch[, c("x", "y")])$lyr1 
        # the naming doesn't quite work
      )
  })
)

survey_sites = survey_sites |>
  select(season, climvar, chelsa, forms) |>
  mutate(
    forms = as.character(forms)
  ) |>
  unnest(
    cols = chelsa
  )

# save measures at survey sites
write_csv(
  survey_sites,
  file = "data/output/data_gam_compare_survey_sites.csv"
)
```
