==> scripts/03_climate_correction_layer.Rmd <==
---
editor_options: 
  chunk_output_type: console
---

---
editor_options: 
  chunk_output_type: console
---

# GAM error in predicting climate

```{r}
library(tidyverse)
library(ggplot2)
library(colorspace)
library(patchwork)

library(terra)
library(stars)
```

## Load data

```{r}
error_data = read_csv("data/output/data_gam_comparison_mae.csv")
survey_clim = read_csv("data/output/data_gam_compare_survey_sites.csv")

# load raster
# load("data/output/model_pred_climate.Rds")
```

## Plot model error

```{r}
# nest data
error_data = nest(
  error_data,
  data = !c("climvar")
)

forms = distinct(survey_clim, form = forms)

# plot data
plots = map2(
  error_data$data, error_data$climvar,
  function(df, cl) {
    ggplot(df)+
      geom_boxplot(
        aes(
          forms, mae
        ),
        width = 0.5
      )+
      geom_text(
        data = forms,
        aes(
          form, 1,
          label = form
        ),
        angle = 90,
        hjust = "inward",
        nudge_x = -0.5,
        col = "steelblue",
        alpha = 0.6,
        fontface = "italic"
      )+
      scale_y_log10()+
      facet_wrap(
        ~season, labeller = label_both
      )+
      theme_grey(base_size = 10)+
      theme(
        strip.background = element_blank(),
        axis.text.y = element_text(
          angle = 90,
          hjust = 0.5
        ),
        axis.text.x = element_blank(),
        panel.border = element_rect(
          fill = NA, colour = "black"
        )
      )+
      labs(
        x = "GAM formula",
        y = "Mean absolute error",
        title = ifelse(
          cl == "ppt", 
          "Precipation",
          "Mean temperature"
        )
      )
  })

plots = wrap_plots(
  plots, ncol = 1
) +
  plot_annotation(
    tag_levels = "A"
  ) &
  theme(
    plot.tag = element_text(
      face = "bold"
    )
  )

ggsave(
  plots, filename = "figures/fig_compare_models_general.png",
  width = 6,
  height = 8
)
```

## Compare climate predictions at survey sites

```{r}
survey_clim = group_by(survey_clim, climvar, season) |>
  nest()

plots_survey_pts = Map(
  survey_clim$climvar, survey_clim$data, survey_clim$season,
  f = function(cvar, df, season) {
    ggplot(
      df
    )+
      geom_abline(
        slope = 1
      )+
      geom_point(
        aes(bioclim_val, pred_val),
        shape = 1,
        alpha = 0.8
      )+
      scale_y_continuous(
        # trans = "sqrt"
        # trans = ggallin::ssqrt_trans
      )+
      facet_grid(
        ~ forms, scales = "free",
        labeller = labeller(
          .multi_line = F,
          season = label_both
        )
      )+
      theme_grey(
        base_size = 8
      )+
      theme(
        legend.position = "top",
        panel.border = element_rect(
          fill = NA, colour = "black"
        )
      )+
      labs(
        colur = "Model",
        title = ifelse(
          cvar == "ppt", 
          glue::glue("Precipation; season: {season}"),
          glue::glue("Mean temperature; season: {season}")
        )
      )
  }
)

plots_survey_pts = wrap_plots(
  plots_survey_pts, ncol = 1
) +
  plot_annotation(
    tag_levels = "A"
  )

ggsave(
  plots_survey_pts,
  filename = "figures/fig_survey_site_model_comparison.png",
  height = 10, width = 6
)
```

## Plot GAM prediction residuals

### Load predicted data

```{r}
# load data
gam_validation_pred = terra::rast("data/output/gam_pred_valid_avg.tif") |>
  as.list()
gam_validation_id = read_csv("data/output/gam_model_formulas.csv")

# select and save
gam_validation_pred = gam_validation_pred[gam_validation_id$climvar %in% 
                                            c("ppt", "t_mean")]
gam_validation_id = filter(gam_validation_id, climvar %in% c("ppt", "t_mean"))

# link -- could have filtered here too -- oh well
gam_validation_data = mutate(gam_validation_id, prediction = gam_validation_pred)
```

### Load BIOCLIM data

```{r}
chelsa_temp = terra::rast("data/output/chelsa_temp_stack.tif") |>
  as.list()
chelsa_ppt = terra::rast("data/output/chelsa_ppt_stack.tif") |>
  as.list()

# make data
chelsa_data =
  crossing(
  season = c("rainy", "dry"),
  climvar = c("t_mean", "ppt")
) |>
  arrange(desc(climvar), desc(season)) |>
  mutate(
    chelsa_rast = append(chelsa_temp, chelsa_ppt)
  )
```

### Get prediction as fraction of BIOCLIM layers

```{r}
# link prediction and residual
gam_validation_data = gam_validation_data |>
  left_join(chelsa_data)

gam_validation_data = mutate(
  gam_validation_data,
  residual = map2(chelsa_rast, prediction, function(ch, pr) {
    pr = terra::resample(pr, ch) # resampling required
    pr / ch
  })
)
```

### Temperature fractions

```{r}
plots_temp_resid = filter(gam_validation_data, climvar == "t_mean") |>
  select(-prediction, -chelsa_rast)

plots_temp_resid = pmap(
  plots_temp_resid,
  .f = function(season, climvar, forms, residual) {
    residual = st_as_stars(residual)
    ggplot()+
      geom_stars(
        data = residual
      )+
      scale_fill_continuous_diverging(
        palette = "Blue-Red 3",
        mid = 1,
        na.value = "transparent",
        name = glue::glue("Prediction / BIOCLIM 
                          {climvar}"),
        limits = c(0.5, 2),
        labels = scales::percent,
        trans = ggallin::ssqrt_trans
      )+
      theme_test(base_size = 6)+
      theme(
        legend.position = "right",
        legend.key.height = unit(10, "mm"),
        legend.key.width = unit(2, "mm"),
        axis.title = element_blank()
      )+
      coord_sf(
        expand = F
      )+
      labs(
        title = glue::glue("variable: {climvar} season: {season} 
                           model: {forms}")
      )
  }
) |> wrap_plots(
  guides = "collect"
) &
  theme(
    legend.position = "right"
  )

plots_temp_resid[[1]]
```

### Precipitation fractions

```{r}
plots_ppt_resid = filter(gam_validation_data, climvar == "ppt") |>
  select(-prediction, -chelsa_rast)
plots_ppt_resid = pmap(
  plots_ppt_resid,
  .f = function(season, climvar, forms, residual) {
    residual = st_as_stars(residual)
    ggplot()+
      geom_stars(
        data = residual
      )+
      scale_fill_continuous_diverging(
        palette = "Vik",
        rev = T,
        mid = 1,
        na.value = "transparent",
        name = glue::glue("Prediction / BIOCLIM 
                          {climvar}"),
        limits = c(0.1, 5.5),
        labels = scales::percent,
        breaks = c(0.01, 0.5, seq(0.0, 5.5, 1))
      )+
      theme_test(base_size = 6)+
      theme(
        legend.position = "right",
        legend.key.height = unit(10, "mm"),
        legend.key.width = unit(2, "mm"),
        axis.title = element_blank()
      )+
      coord_sf(
        expand = F
      )+
      labs(
        title = glue::glue("variable: {climvar} season: {season} 
                           model: {forms}")
      )
  }
) |> wrap_plots(
  guides = "collect"
) &
  theme(
    legend.position = "right"
  )

plots_ppt_resid[[1]]
```

```{r}
# save residual plots
ggsave(
  plots_temp_resid,
  filename = "figures/fig_temp_resid.png",
  width = 9, height = 7
)

ggsave(
  plots_ppt_resid,
  filename = "figures/fig_ppt_resid.png",
  width = 9, height = 7
)
```

## Save climate correction layers

Looking at the GAM predictions as fractions of the BIOCLIM layers, we can choose model formulas for each season and each variable that lead to predictions that are closest to the real BIOCLIM values. We choose formulas on the basis of observed deviation from the true value, as well as spatial contiguity of deviations, basically, are nearby areas similarly different from true values. For example, there is an odd north east regional deviation for rainfall in the wet season for the forumla $\text{ppt} ~ s(\text{elevation}, k = 3) + s(\text{distance to coast, latitude}, k = 5)$, so we prefer to chose another formula.

This means we pick the simple $\text{temp} ~ s(\text{elevation}, k = 3)$ formula for mean monthly temperature, in both dry and wet seasons, and the $\text{ppt} ~ s(\text{elevation}, k = 3) + \text{distance to coast} + \text{latitude}$ formula for total monthly rainfall in both dry and wet seasons.

We then save the fraction, BIOCLIM / prediction, as a correction layer --- one layer per season and variable.
This allows us to fit GAMs to chunks of historical climate data, using physical predictors as covariates, and to then correct the resulting spatial prediction using the correction layer.
Hence the 'true' historical value of a climate variable is $\text{GAM prediction} \times \text{correction factor}$, where the correction factor is the cell-specific value from the correction layers.

### Seasonal mean temperature correction layer

```{r}
# subset validation dataset
temp_correction_layer = gam_validation_data %>% 
  filter(
    climvar == "t_mean",
    forms == "value~s(elev, k = 3)"
  )

# now get the correction layer
temp_correction_layer = mutate(
  temp_correction_layer,
  correction_layer = map2(
    chelsa_rast, prediction, function(ch, pr) {
      pr = terra::resample(pr, ch) # resampling required
      ch / pr
    }
  )
)

# make the correction layer a single raster stack object and name correctly
temp_correction_layer = Reduce(
  temp_correction_layer$correction_layer, f = c
)

# set names for layers
names(temp_correction_layer) = c(
  "correction_layer_temp_dry",
  "correction_layer_temp_wet"
)

# save the layer
terra::writeRaster(
  temp_correction_layer,
  filename = "data/output/raster_correction_layers_temp.tif"
)
```

### Seasonal total rainfall correction layer

```{r}
# subset validation dataset
ppt_correction_layer = gam_validation_data %>% 
  filter(
    climvar == "ppt",
    forms == "value~s(elev, k = 3) + coast + lat"
  )

# now get the correction layer
ppt_correction_layer = mutate(
  ppt_correction_layer,
  correction_layer = map2(
    chelsa_rast, prediction, function(ch, pr) {
      pr = terra::resample(pr, ch) # resampling required
      ch / pr
    }
  )
)

# make the correction layer a single raster stack object and name correctly
ppt_correction_layer = Reduce(
  ppt_correction_layer$correction_layer, f = c
)

# set names for layers
names(ppt_correction_layer) = c(
  "correction_layer_ppt_dry",
  "correction_layer_ppt_wet"
)

# save the layer
terra::writeRaster(
  ppt_correction_layer,
  filename = "data/output/raster_correction_layers_ppt.tif"
)
```
