==> scripts/04_recontstruct_climate.Rmd <==
---
editor_options: 
  chunk_output_type: console
---

# Reconstructing climate data

## Prepare weather and spatial data

These steps are identical to those in `run_clim_gams`, except that we shall run only a single model per season and climate variable, but for multiple time chunks, of about 30 years.

```{r}
# for data
library(tidyverse)
library(glue)

# for gam
library(mgcv)

# for rasters
library(terra)
```

Load data prepared for the GAM, and the stack predictor raster.

```{r}
# read data
data = read_csv("data/output/data_for_gam.csv")
stack = terra::rast("data/output/raster_gam_stack.tif")
```

Prepare data covariates and raster.

```{r}
# prepare data for gam fitting
data = mutate(
  data,
  
  # divide distance by 1000 for km
  coast = coast / 1000,
  
  # add some error to all physical variables
  # hopefully prevents model issues
  coast = coast + rnorm(length(coast), 0.1, sd = 0.01),
  elev = elev + rnorm(length(coast), 1, sd = 0.01),
  lat = lat + rnorm(length(lat), 0, sd = 0.01)
)

# set distance to coast raster in km
values(stack[["coast"]]) = values(stack[["coast"]]) / 1000
```

Pivot longer to separate by variable.

```{r}
# pivot longer
data = pivot_longer(
    data,
    cols = c("ppt", "t_mean", "t_sd"),
    names_to = "climvar"
)
data = dplyr::select(data, !month)

# remove t_sd
data = filter(
  data, climvar != "t_sd"
)
```

Split data by variable, season, and year bin (already assigned in `prep_gam_predictors`.

```{r}
data = nest(
  data, data = !c(climvar, season, year_bin)
)
```

## Fit chosen GAMs

Specify GAM formulas for temperature and precipitation.

```{r}
# model formula for temp
form_temp = "value ~ s(elev, k = 3)"

# model formula for ppt
form_ppt = "value ~ s(elev, k = 3) + coast + lat"
```

Assign formula to variable.

```{r}
# assign formula as a character for now
data = mutate(
  data,
  form = if_else(
    climvar == "t_mean", form_temp, form_ppt
  )
)
```

Fit GAMs.

```{r}
# fit models
data = mutate(
  data,
  mod = map2(data, form, function(df, form) {
    gam(
      formula = as.formula(form),
      data = df
    )
  })
)
```

Save model data.

```{r}
save(data, file = "data/output/model_reconstruction_climate.Rds")
```

## Predict over static environment

Get predictions

```{r}
# save model predictions
model_pred = map(
  data$mod, function(g) {
    predict(stack, g, type = "response") # order matters here
  }
)
```

```{r}
# assign predicted rasters as list column object
data = mutate(
  data,
  pred = model_pred
)
```

```{r}
# subset columns
data_pred = ungroup(data) %>% 
  select(season, year_bin, climvar, pred)
```

## Apply correction layers

Read in correction layers.

```{r}
# read in correction layers
temp_correction = terra::rast(
  "data/output/raster_correction_layers_temp.tif"
)

ppt_correction = terra::rast(
  "data/output/raster_correction_layers_ppt.tif"
)
```

Link correction layers with predictions.

```{r}
# assign correction layer
data_pred = mutate(
  data_pred,
  correction = map2(
    season, climvar, function(season, climvar) {
      if(season == "dry" & climvar == "ppt") {
        ppt_correction[["correction_layer_ppt_dry"]]
      } else if(season == "dry" & climvar == "temp") {
        temp_correction[["correction_layer_temp_dry"]]
      } else if(season == "wet" & climvar == "ppt") {
        ppt_correction[["correction_layer_ppt_wet"]]
      } else temp_correction[["correction_layer_temp_wet"]]
    }
  )
)

# make correction to prediction
data_pred = mutate(
  data_pred,
  corrected = map2(
    correction, pred, function(ch, pr) {
      pr = terra::resample(pr, ch) # resampling required
      ch * pr
    }
  )
)

# assign names to rasters
data_pred = mutate(
  data_pred, 
  names = glue("{climvar}_{year_bin}_{season}")
)

names(data_pred$corrected) = data_pred$names
```

```{r}
data_reconstructed = data_pred %>% 
  group_by(
    season, climvar
  ) %>% 
  summarise(
    corrected = list(Reduce(corrected, f = c))
  )

# save each as a tif
pwalk(
  data_reconstructed,
  function(season, climvar, corrected) {
    
    names(corrected) = glue("{climvar}_{season}_{seq(1870, 2018, 10)}")
    
    terra::writeRaster(
      corrected,
      glue("data/output/raster_reconstructed_{climvar}_{season}.tif")
    )
  }
)
```
